@@ -19,8 +19,8 @@
     "@nuxtjs/vuetify": "^1.12.3",
     "@pinia/nuxt": "^0.1.9",
     "@vue/apollo-composable": "^4.0.0-alpha.17",
-    "@vueuse/core": "^8.5.0",
-    "@vueuse/integrations": "^8.5.0",
+    "@vueuse/core": "^8.6.0",
+    "@vueuse/integrations": "^8.6.0",
     "accounting": "^0.4.1",
     "apexcharts": "^3.35.3",
     "detect-browser": "^5.3.0",
@@ -41,13 +41,13 @@
     "@babel/plugin-proposal-optional-chaining": "^7.17.12",
     "@graphql-codegen/cli": "^2.6.2",
     "@graphql-codegen/schema-ast": "^2.4.1",
-    "@graphql-codegen/typescript": "^2.4.11",
-    "@graphql-codegen/typescript-operations": "^2.4.0",
-    "@graphql-codegen/typescript-vue-apollo": "^3.2.12",
+    "@graphql-codegen/typescript": "^2.5.0",
+    "@graphql-codegen/typescript-operations": "^2.4.1",
+    "@graphql-codegen/typescript-vue-apollo": "^3.2.13",
     "@mdi/font": "^6.7.96",
     "@nuxt/bridge": "npm:@nuxt/bridge-edge",
     "@nuxt/types": "^2.15.8",
-    "@nuxtjs/color-mode": "^3.0.3",
+    "@nuxtjs/color-mode": "^3.1.3",
     "@nuxtjs/eslint-config-typescript": "^10.0.0",
     "@tiptap/core": "^2.0.0-beta.176",
     "@tiptap/extension-bold": "^2.0.0-beta.26",
@@ -78,17 +78,15 @@
     "@tiptap/extension-underline": "^2.0.0-beta.23",
     "@tiptap/starter-kit": "^2.0.0-beta.185",
     "@tiptap/vue-2": "^2.0.0-beta.79",
-    "@typescript-eslint/eslint-plugin": "^5.27.0",
-    "@typescript-eslint/parser": "^5.27.0",
-    "core-js": "^3.22.7",
-    "eslint": "^8.16.0",
+    "@typescript-eslint/eslint-plugin": "^5.27.1",
+    "@typescript-eslint/parser": "^5.27.1",
+    "core-js": "^3.22.8",
+    "eslint": "^8.17.0",
     "eslint-import-resolver-typescript": "^2.7.1",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-nuxt": "^3.2.0",
     "pug": "^3.0.2",
     "pug-plain-loader": "^1.1.0",
-    "typescript": "^4.7.2",
-    "vue-class-component": "^7.2.6",
-    "vue-property-decorator": "^9.1.2"
+    "typescript": "^4.7.3"
   }
 }
