@@ -1,72 +1,70 @@
 <template lang="pug">
   thead
     tr
-      th(:style="{ width: `${rowIndexColumnWidth}px` }" @click="selectAll")
-        .grid__header-content
-          .grid__select-all(:class="{ 'grid__select-all_selected': allSelected }")
       th(
-        v-for="column in columns"
-        :key="column.id"
-        :style="column.style"
-        @mouseenter="mouseenterColumnIndex(column)"
-        @mousemove="mousemoveColumnIndex($event, column)"
-        @mouseleave="mouseleaveColumnIndex"
-        @mousedown="mousedownColumnIndex($event, column)"
-        @mouseup="mouseupColumnIndex"
+        :class="{ 'grid__header_all_selected': allCellsSelected }"
+        :style="{ width: `${rowNameColumnWidth}px` }"
+        @click="selectAllCells"
       )
-        grid-column-control(v-slot="{ on, attrs }" :column="column")
-          div(
-            v-bind="attrs"
-            @contextmenu.prevent="on.click"
-            :class="getHeaderContentClasses(column)"
-          ) {{ column.position }}
+        div
+      th(
+        v-for="column in activeSheet.columns"
+        :key="column.id"
+        :class="getHeaderClass(column)"
+        :style="{ 'width': `${getColumnWidth(column)}px` }"
+        @mouseenter="mouseenterColumnName(column)"
+        @mousemove="mousemoveColumnName(column, $event)"
+        @mouseleave="mouseleaveColumnName"
+        @mousedown="mousedownColumnName(column, $event)"
+        @mouseup="mouseupColumnName"
+      )
+        grid-column-control(v-slot="{ on, attrs }" :column="column" :get-column-width="getColumnWidth")
+          div(v-bind="attrs" @contextmenu.prevent="on.click") {{ column.name }}
 </template>
 
 <script lang="ts">
-import type { PropType } from '#app'
-import { defineComponent } from '#app'
-import { BuildColumnType, BoundaryRowCell } from '~/types/grid-types'
+import { PropType, Ref } from '#app'
+import { SheetType, ColumnDimensionType } from '~/types/graphql'
+import { ResizingType } from '~/types/grid'
 import GridColumnControl from '~/components/dcis/grid/controls/GridColumnControl.vue'
 
 export default defineComponent({
   components: { GridColumnControl },
   props: {
-    rowIndexColumnWidth: { type: Number, required: true },
-    columns: { type: Array as PropType<BuildColumnType[]>, required: true },
-    selectionColumns: { type: Array as PropType<number[]>, required: true },
-    selectedBoundaryRowCells: { type: Array as PropType<BoundaryRowCell[]>, required: true },
-    allSelected: { type: Boolean, required: true },
-    mouseenterColumnIndex: {
-      type: Function as PropType<(column: BuildColumnType) => void>,
+    rowNameColumnWidth: { type: Number, required: true },
+    resizingColumn: { type: Object as PropType<ResizingType<ColumnDimensionType>>, default: null },
+    getColumnWidth: { type: Function as PropType<(column: ColumnDimensionType) => number>, required: true },
+    selectedColumnPositions: { type: Array as PropType<number[]>, required: true },
+    boundarySelectedColumnsPositions: { type: Array as PropType<number[]>, required: true },
+    allCellsSelected: { type: Boolean, required: true },
+    mouseenterColumnName: {
+      type: Function as PropType<(column: ColumnDimensionType) => void>,
       required: true
     },
-    mousemoveColumnIndex: {
-      type: Function as PropType<(event: MouseEvent, column: BuildColumnType) => void>,
+    mousemoveColumnName: {
+      type: Function as PropType<(column: ColumnDimensionType, event: MouseEvent) => void>,
       required: true
     },
-    mouseleaveColumnIndex: { type: Function as PropType<() => void>, required: true },
-    mousedownColumnIndex: {
-      type: Function as PropType<(event: MouseEvent, column: BuildColumnType) => void>,
+    mouseleaveColumnName: { type: Function as PropType<() => void>, required: true },
+    mousedownColumnName: {
+      type: Function as PropType<(column: ColumnDimensionType, event: MouseEvent) => void>,
       required: true
     },
-    mouseupColumnIndex: { type: Function as PropType<() => void>, required: true },
-    selectAll: { type: Function as PropType<() => void>, required: true }
+    mouseupColumnName: { type: Function as PropType<() => void>, required: true },
+    selectAllCells: { type: Function as PropType<() => void>, required: true }
   },
   setup (props) {
-    const getHeaderContentClasses = (column: BuildColumnType): (string | Record<string, boolean>)[] => {
-      return [
-        'grid__header-content',
-        { 'grid__header-content_selected': props.selectionColumns.includes(column.index) },
-        {
-          'grid__header-content_neighbor-selected':
-            !!props.selectedBoundaryRowCells.find(boundaryCell =>
-              boundaryCell.columns.find(boundaryRowCell => boundaryRowCell.id === column.id))
-        }
-      ]
+    const activeSheet = inject<Ref<SheetType>>('activeSheet')
+
+    const getHeaderClass = (column: ColumnDimensionType): Record<string, boolean> => {
+      return {
+        grid__header_selected: props.selectedColumnPositions.includes(column.index),
+        'grid__header_boundary-selected': props.boundarySelectedColumnsPositions.includes(column.index),
+        grid__header_hover: !props.resizingColumn
+      }
     }
-    return {
-      getHeaderContentClasses
-    }
+
+    return { activeSheet, getHeaderClass }
   }
 })
 </script>
