@@ -1,12 +1,12 @@
 <template lang="pug">
   input(
     v-focus
+    :value="value"
+    type="text"
     @keyup.esc="$emit('cancel')"
     @keyup.enter="$event.target.blur()"
     @keyup.tab.stop="$event.target.blur()"
     @blur="setValue($event.target.value)"
-    :value="value"
-    type="text"
   )
 </template>
 
