@@ -1,12 +1,12 @@
 <template lang="pug">
   input(
     v-focus
+    :value="value"
+    type="number"
     @keyup.esc="$emit('cancel')"
     @keyup.enter="$event.target.blur()"
     @keyup.tab.stop="$event.target.blur()"
     @blur="$emit('set-value', $event.target.value)"
-    :value="value"
-    type="number"
   )
 </template>
 
