@@ -3,7 +3,7 @@
     template(#activator="{ on }")
       div(v-on="on") {{ value }}
     v-card
-      v-card-title {{ t('dcis.cells.gridCellText.title') }}
+      v-card-title {{ t('dcis.grid.changeValue') }}
         v-spacer
         v-btn(@click="cancel" icon)
           v-icon mdi-close
@@ -11,7 +11,7 @@
         v-textarea(v-model="rawValue" auto-grow autofocus)
       v-card-actions
         v-spacer
-        v-btn(@click="setValue" color="primary") {{ $t('save') }}
+        v-btn(@click="setValue" color="primary") {{ t('save') }}
 </template>
 
 <script lang="ts">
