@@ -3,30 +3,36 @@
     template(#activator="{ on, attrs }")
       slot(:on="on" :attrs="attrs")
     v-list(dense)
-      column-settings(@close="active = false" :column="column")
+      grid-column-settings(
+        :column="column"
+        :get-column-width="getColumnWidth"
+        @close="active = false"
+      )
         template(#activator="{ on }")
           v-list-item(v-on="on")
             v-list-item-icon
               v-icon mdi-cog
             v-list-item-content
-              v-list-item-title Свойства
+              v-list-item-title {{ t('dcis.grid.columnControl.properties') }}
 </template>
 
 <script lang="ts">
-import type { PropType } from '#app'
-import { defineComponent, ref } from '#app'
-import { BuildColumnType } from '~/types/grid-types'
-import ColumnSettings from '~/components/dcis/documents/ColumnSettings.vue'
+import { PropType } from '#app'
+import { ColumnDimensionType } from '~/types/graphql'
+import GridColumnSettings from '~/components/dcis/grid/settings/GridColumnSettings.vue'
 
 export default defineComponent({
-  components: { ColumnSettings },
+  components: { GridColumnSettings },
   props: {
-    column: { type: Object as PropType<BuildColumnType>, required: true }
+    column: { type: Object as PropType<ColumnDimensionType>, required: true },
+    getColumnWidth: { type: Function as PropType<(column: ColumnDimensionType) => number>, required: true }
   },
   setup () {
+    const { t } = useI18n()
+
     const active = ref<boolean>(false)
 
-    return { active }
+    return { t, active }
   }
 })
 </script>
