@@ -1,108 +1,81 @@
 <template lang="pug">
-  v-dialog(v-model="drawer" width="600")
+  mutation-modal-form(
+    @done="changePageAvatarDone"
+    :mutation="require('~/gql/pages/mutations/page/change_page_avatar.graphql')"
+    :variables="{ pageId: page.id, avatar }"
+    :header="$t('pages.page.changeAvatar.header')"
+    :button-text="buttonText"
+    mutation-name="changePageAvatar"
+    i18n-path="pages.page.changeAvatar"
+  )
     template(#activator="{ on }")
       slot(:on="on")
-    apollo-mutation(
-      :mutation="require('~/gql/pages/mutations/page/change_page_avatar.graphql')"
-      :variables="{ pageId: page.id, avatar }"
-      @done="changePageAvatarDone"
-      v-slot="{ mutate, loading, error }"
-    )
-      v-card
-        v-card-title {{ $t('pages.page.changeAvatar.header') }}
-          v-spacer
-          v-btn(@click="close" icon)
-            v-icon mdi-close
-        v-card-subtitle {{ subtitle }}
-        v-card-text
-          v-alert(type="error" :value="!!error" dismissible) {{ error }}
-          v-img.mb-4(v-if="avatar || page.avatar" :src="avatarSrc" height="450")
-          v-file-input(
-            v-model="avatar"
-            :label="$t('pages.page.changeAvatar.avatar')"
-            prepend-icon="mdi-camera"
-            show-size
-          )
-        v-card-actions
-          v-spacer
-          v-btn(
-            @click="mutate()"
-            :disabled="!(avatar || page.avatar)"
-            :loading="loading"
-            color="primary"
-          ) {{ buttonText }}
+    template(#form)
+      v-img.mb-4(v-if="avatar || page.avatar" :src="avatarSrc" height="450")
+      v-file-input(
+        v-model="avatar"
+        :label="$t('pages.page.changeAvatar.avatar')"
+        prepend-icon="mdi-camera"
+        show-size
+        @change="setAvatarSrc"
+      )
+    template(#actions="{ invalid, loading, buttonText, setFormErrors, setError, setSuccess }")
+      v-spacer
+      v-btn(
+        :disabled="invalid || !(avatar || page.avatar)"
+        :loading="loading"
+        type="submit"
+        color="primary"
+      ) {{ buttonText }}
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from 'vue-property-decorator'
-import { PageType, ChangePageAvatarMutation } from '~/types/graphql'
+import { defineComponent, PropType, ref, computed } from '#app'
+import { useI18n } from '~/composables'
+import { PageType } from '~/types/graphql'
+import MutationModalForm from '~/components/common/forms/MutationModalForm.vue'
 
-@Component<ChangePageAvatar>({
-  computed: {
-    serverAvatarName () {
-      return this.page.avatar ? this.page.avatar.match(/[^/]+$/)![0] : null
-    },
-    subtitle () {
-      return this.page.avatar
-        ? this.$t('pages.page.changeAvatar.serverAvatar', { name: this.serverAvatarName })
-        : this.$t('pages.page.changeAvatar.noServerAvatar')
-    },
-    avatarSrc () {
-      if (this.localAvatarSrc) {
-        return this.localAvatarSrc
+export default defineComponent({
+  components: { MutationModalForm },
+  props: { page: { required: true, type: Object as PropType<PageType> } },
+  setup (props, { emit }) {
+    const { t } = useI18n()
+
+    const avatar = ref<File | null>(null)
+    const localAvatarSrc = ref<string | null>(null)
+
+    const avatarSrc = computed(() => {
+      if (localAvatarSrc.value) {
+        return localAvatarSrc.value
       }
-      return `/${this.page.avatar}`
-    },
-    buttonText () {
-      return this.avatar != null
-        ? this.$t('pages.page.changeAvatar.change')
-        : this.$t('pages.page.changeAvatar.delete')
-    }
-  },
-  subscriptions () {
-    const avatarWatch = this.$watchAsObservable('avatar')
-    avatarWatch.subscribe(({ newValue }: { newValue: File | null }) => {
-      this.setAvatarSrc(newValue)
+      return `/${props.page.avatar}`
     })
-    return {
-      avatarWatch
+
+    const buttonText = computed(() => {
+      return avatar.value
+        ? t('pages.page.changeAvatar.change')
+        : t('pages.page.changeAvatar.delete')
+    })
+
+    const setAvatarSrc = async () => {
+      if (avatar.value) {
+        const fr = new FileReader()
+        const result = new Promise<string>((resolve) => {
+          fr.onload = () => resolve(fr.result as string)
+        })
+        fr.readAsDataURL(avatar.value)
+        localAvatarSrc.value = await result
+      } else {
+        localAvatarSrc.value = null
+      }
     }
+
+    const changePageAvatarDone = () => {
+      avatar.value = null
+      emit('close')
+    }
+
+    return { avatar, avatarSrc, setAvatarSrc, buttonText, changePageAvatarDone }
   }
 })
-export default class ChangePageAvatar extends Vue {
-  @Prop({ required: true, type: Object }) readonly page!: PageType
-
-  drawer: boolean = false
-  avatar: File | null = null
-  localAvatarSrc: string | null = null
-  serverAvatarName!: string | null
-  subtitle!: string
-  avatarSrc!: string | null | undefined
-  buttonText!: string
-
-  async setAvatarSrc (avatar: File | null) {
-    if (avatar != null) {
-      const fr = new FileReader()
-      const result = new Promise<string>((resolve) => {
-        fr.onload = () => resolve(fr.result as string)
-      })
-      fr.readAsDataURL(avatar)
-      this.localAvatarSrc = await result
-    } else {
-      this.localAvatarSrc = null
-    }
-  }
-
-  changePageAvatarDone ({ data: { changePageAvatar: { success } } }: { data: ChangePageAvatarMutation }) {
-    if (success) {
-      this.close()
-    }
-  }
-
-  close () {
-    this.drawer = false
-    this.avatar = null
-    this.$emit('close')
-  }
-}
 </script>
