@@ -1,4 +1,5 @@
 import Vue, { ComponentOptions } from 'vue'
+import { Ref } from '#app'
 import { SettingType, UserType } from '~/types/graphql'
 
 declare module 'vue/types' {
@@ -27,27 +28,7 @@ declare module 'vue/types/vue' {
       textLength(str: string, length: number): string,
       timeHM(str: string): string
     }
-  }
-}
-
-declare module 'vuex/types/index' {
-  // this.$myInjectedFunction inside Vuex stores
-  interface Store<S> {
-    $getNowDate(): string
-    $fromGlobalId(globalId: string): { type: string, id: string }
-    $getSettingValue(key: string): string | null
-    $cursor(n: number): string
-    $toGlobalId(type: string, id: number): string
-    $getUserName(user: UserType, showSirName?: boolean): string
-    $getUserFullName(user: UserType, showSirName?: boolean): string
-    $filters: {
-      money(str: string): string,
-      date(str: string): string,
-      dateTimeHM(str: string): string,
-      basename(str: string): string,
-      textLength(str: string, length: number): string,
-      timeHM(str: string): string
-    }
+    $colorMode: Ref<string>
   }
 }
 
