@@ -1,4 +1,4 @@
-from openpyexcel.utils.cell import get_column_letter
+from openpyxl.utils.cell import get_column_letter
 from django.db import models
 
 from apps.core.models import User
@@ -68,7 +68,7 @@ class Style(models.Model):
     size = models.PositiveIntegerField(default=12, help_text='Размер шрифта')
     strong = models.BooleanField(default=False, help_text='Жирный шрифт')
     italic = models.BooleanField(default=False, help_text='Курсив')
-    strike = models.BooleanField(default=None, null=True, help_text='Зачеркнутый')
+    strike = models.BooleanField(default=False, help_text='Зачеркнутый')
     underline = models.CharField(
         max_length=20,
         default=None,
@@ -150,7 +150,10 @@ class ColumnDimension(KindCell, models.Model):
     index = models.PositiveIntegerField(help_text='Индекс колонки')
     width = models.PositiveIntegerField(null=True, help_text='Ширина колонки')
     fixed = models.BooleanField(default=False, help_text='Фиксация колонки')
-    hidden = models.BooleanField(default=False, help_text='Скрытое поле')
+    hidden = models.BooleanField(default=False, help_text='Скрытие колонки')
+
+    created_at = models.DateTimeField(auto_now_add=True, help_text='Дата добавления')
+    updated_at = models.DateTimeField(auto_now=True, help_text='Дата обновления')
 
     sheet = models.ForeignKey(Sheet, on_delete=models.CASCADE, help_text='Лист')
     user = models.ForeignKey(User, null=True, on_delete=models.SET_NULL, help_text='Пользователь')
@@ -187,7 +190,9 @@ class RowDimension(models.Model):
     )
 
     index = models.PositiveIntegerField(help_text='Индекс строки')
-    height = models.PositiveIntegerField(null=True, help_text='Высота колонки')
+    height = models.PositiveIntegerField(null=True, help_text='Высота строки')
+    fixed = models.BooleanField(default=False, help_text='Фиксация строки')
+    hidden = models.BooleanField(default=False, help_text='Скрытие строки')
 
     dynamic = models.BooleanField(default=False, help_text='Динамическая ли строка')
     object_id = models.PositiveIntegerField(null=True)
@@ -222,10 +227,6 @@ class RowDimension(models.Model):
 
     class Meta:
         ordering = ('index', 'id',)
-        # indexes = [
-        #     models.Index(fields=['document', 'parent']),
-        #     models.Index(fields=['document', 'parent', 'object_id'])
-        # ]
 
 
 class Cell(Style, KindCell, models.Model):
@@ -312,8 +313,8 @@ class MergedCell(models.Model):
     @property
     def cells(self) -> list[str]:
         not_cell: list[str] = []
-        for col in range(self.min_col, self.max_col + 1):
-            for row in range(self.min_row, self.max_row + 1):
+        for row in range(self.min_row, self.max_row + 1):
+            for col in range(self.min_col, self.max_col + 1):
                 if col == 1 and row == 1:
                     continue
                 not_cell.append(f'{get_column_letter(col)}{row}')
