@@ -1,4 +1,4 @@
-from typing import Any, Optional
+from typing import Any
 
 import graphene
 from devind_core.schema import FileType
@@ -8,119 +8,213 @@ from devind_helpers.permissions import IsAuthenticated
 from devind_helpers.schema.mutations import BaseMutation
 from devind_helpers.schema.types import ErrorFieldType
 from django.core.files.uploadedfile import InMemoryUploadedFile
-from django.db import transaction
-from django.db.models import F
-from graphene_django_cud.mutations import DjangoUpdateMutation
 from graphene_file_upload.scalars import Upload
 from graphql import ResolveInfo
 from graphql_relay import from_global_id
+from stringcase import snakecase
 
-from apps.dcis.helpers import DjangoCudBaseMutation
-from apps.dcis.models import Cell, ColumnDimension, Document, RowDimension, Sheet, Value
-from apps.dcis.schema.types import CellType, ColumnDimensionType, MergedCellType, RowDimensionType, ValueType
-from apps.dcis.services.cell import change_cell_kind, check_cell_options
-from apps.dcis.services.value import (
+
+from apps.dcis.models import Cell, ColumnDimension, RowDimension, Sheet, Value
+from apps.dcis.schema.types import SheetType, CellType, ChangedCellOption, GlobalIndicesInputType, RowDimensionType
+from apps.dcis.permissions import ChangeSheet
+from apps.dcis.services.sheet_services import (
+    rename_sheet,
+    CheckCellOptions,
+    add_row_dimension,
+    change_cells_option,
+    change_column_dimension,
+    change_row_dimension,
     create_file_value_archive,
+    delete_row_dimension,
     get_file_value_files,
     update_or_create_file_value,
     update_or_create_value,
-    updates_values_by_cell_kind_change,
 )
 
 
-class AddRowDimensionMutation(BaseMutation):
-    """Вставка строк.
+class RenameSheetMutation(BaseMutation):
+    """Изменение названия листа.
 
-    После добавления строки бы то не было, строка приобретает новый индекс,
-    соответственно, все строки после вставленной строки должны увеличить свой индекс на единицу.
+    Во время мутации изменяем только формулы и ничего не пересчитываем.
     """
 
     class Input:
-        document_id = graphene.ID(required=True, description='Идентификатор документа')
-        row_id = graphene.Int(required=True, description='Идентификатор строки')
-        position = graphene.String(required=True, description='Позиция вставки')
+        sheet_id = graphene.ID(required=True, description='Идентификатор листа')
+        name = graphene.String(required=True, description='Новое название листа')
 
-    row_dimension = graphene.Field(RowDimensionType, required=True, description='Добавленная строка')
-    cells = graphene.List(CellType, required=True, description='Добавленные ячейки')
-    merged_cells = graphene.List(MergedCellType, required=True, description='Объединенные ячейки')
+    sheet = graphene.Field(SheetType, description='Лист')
+    cells = graphene.List(CellType, description='Измененные ячейки')
+
+    @staticmethod
+    @permission_classes((IsAuthenticated, ChangeSheet,))
+    def mutate_and_get_payload(root: Any, info: ResolveInfo, sheet_id: str, name: str):
+        sheet, cells = rename_sheet(get_object_or_404(Sheet, pk=sheet_id), name)
+        return RenameSheetMutation(
+            sheet=sheet,
+            cells=cells
+        )
+
+
+class ChangeColumnDimensionMutation(BaseMutation):
+    """Изменение колонки."""
+
+    class Input:
+        column_dimension_id = graphene.ID(required=True, description='Идентификатор колонки')
+        width = graphene.Int(description='Ширина колонки')
+        fixed = graphene.Boolean(required=True, description='Фиксация колонки')
+        hidden = graphene.Boolean(required=True, description='Скрытие колонки')
+        kind = graphene.String(required=True, description='Тип значения')
+
+    column_dimension_id = graphene.ID(required=True, description='Идентификатор колонки')
+    width = graphene.Int(description='Ширина колонки')
+    fixed = graphene.Boolean(required=True, description='Фиксация колонки')
+    hidden = graphene.Boolean(required=True, description='Скрытие колонки')
+    kind = graphene.String(required=True, description='Тип значения')
+    updated_at = graphene.DateTime(required=True, description='Дата обновления колонки')
 
     @staticmethod
     @permission_classes((IsAuthenticated,))
-    def mutate_and_get_payload(root: Any, info: ResolveInfo, document_id: str, row_id: int, position: str):
-        if position not in ['before', 'after']:
-            return AddRowDimensionMutation(
-                success=False,
-                errors=[ErrorFieldType('position', ['Параметр позиции {position} не из списка: before, after'])]
-            )
-        document: Document = get_object_or_404(Document, pk=from_global_id(document_id)[1])
-        row: RowDimension = get_object_or_404(RowDimension, pk=row_id)
-        insert_index = row.index if position == 'before' else row.index + 1
-        sheet: Sheet = Sheet.objects.get(pk=row.sheet_id)
-        sheet.rowdimension_set.filter(index__gte=insert_index).update(index=F('index') + 1)
-        row_dimension = RowDimension.objects.create(
-            sheet=sheet,
-            index=insert_index,
-            document=document,
-            user=info.context.user
+    def mutate_and_get_payload(
+        root: Any,
+        info: ResolveInfo,
+        column_dimension_id: str,
+        width: int | None,
+        fixed: bool,
+        hidden: bool,
+        kind: str
+    ):
+        column_dimension = change_column_dimension(
+            get_object_or_404(ColumnDimension, pk=column_dimension_id),
+            width=width,
+            fixed=fixed,
+            hidden=hidden,
+            kind=kind
         )
-        cells = [
-            Cell.objects.create(row=row_dimension, column=column, kind=column.kind)
-            for column in sheet.columndimension_set.all()
-        ]
-        sheet.move_merged_cells(insert_index, 1)
+        return ChangeColumnDimensionMutation(
+            column_dimension_id=column_dimension.pk,
+            width=column_dimension.width,
+            fixed=column_dimension.fixed,
+            hidden=column_dimension.hidden,
+            kind=column_dimension.kind,
+            updated_at=column_dimension.updated_at
+        )
+
+
+class AddRowDimensionMutation(BaseMutation):
+    """Добавление строки."""
+
+    class Input:
+        sheet_id = graphene.ID(required=True, description='Идентификатор листа')
+        document_id = graphene.ID(description='Идентификатор документа')
+        parent_id = graphene.ID(description='Идентификатор родительской строки')
+        index = graphene.Int(required=True, description='Индекс вставки')
+        global_index = graphene.Int(required=True, description='Индекс вставки в плоскую структуру')
+        global_indices = graphene.List(
+            graphene.NonNull(GlobalIndicesInputType),
+            required=True,
+            description='Вспомогательные индексы в плоской структуре'
+        )
+
+    row_dimension = graphene.Field(RowDimensionType, required=True, description='Добавленная строка')
+
+    @staticmethod
+    @permission_classes((IsAuthenticated,))
+    def mutate_and_get_payload(
+        root: Any,
+        info: ResolveInfo,
+        document_id: str | None,
+        sheet_id: int,
+        parent_id: str | None,
+        index: int,
+        global_index: int,
+        global_indices: list[GlobalIndicesInputType]
+    ):
         return AddRowDimensionMutation(
-            row_dimension=row_dimension,
-            cells=cells,
-            merged_cells=sheet.mergedcell_set.all()
+            row_dimension=add_row_dimension(
+                user=info.context.user,
+                sheet=get_object_or_404(Sheet, pk=sheet_id),
+                document_id=from_global_id(document_id)[1] if document_id else None,
+                parent_id=int(parent_id) if parent_id else None,
+                index=index,
+                global_index=global_index,
+                global_indices_map={int(i.row_id): i.global_index for i in global_indices}
+            )
+        )
+
+
+class ChangeRowDimensionMutation(BaseMutation):
+    """Изменение строки."""
+
+    class Input:
+        row_dimension_id = graphene.ID(required=True, description='Идентификатор строки')
+        height = graphene.Int(description='Высота строки')
+        fixed = graphene.Boolean(required=True, description='Фиксация строки')
+        hidden = graphene.Boolean(required=True, description='Скрытие строки')
+        dynamic = graphene.Boolean(required=True, description='Динамическая ли строка')
+
+    row_dimension_id = graphene.ID(required=True, description='Идентификатор строки')
+    height = graphene.Int(description='Высота строки')
+    fixed = graphene.Boolean(required=True, description='Фиксация строки')
+    hidden = graphene.Boolean(required=True, description='Скрытие строки')
+    dynamic = graphene.Boolean(required=True, description='Динамическая ли строка')
+    updated_at = graphene.DateTime(required=True, description='Дата обновления строки')
+
+    @staticmethod
+    @permission_classes((IsAuthenticated,))
+    def mutate_and_get_payload(
+        root: Any,
+        info: ResolveInfo,
+        row_dimension_id: str,
+        height: int | None,
+        fixed: bool,
+        hidden: bool,
+        dynamic: bool
+    ):
+        row_dimension = change_row_dimension(
+            get_object_or_404(RowDimension, pk=row_dimension_id),
+            height=height,
+            fixed=fixed,
+            hidden=hidden,
+            dynamic=dynamic
+        )
+        return ChangeRowDimensionMutation(
+            row_dimension_id=row_dimension.pk,
+            height=row_dimension.height,
+            fixed=row_dimension.fixed,
+            hidden=row_dimension.hidden,
+            dynamic=row_dimension.dynamic,
+            updated_at=row_dimension.updated_at
         )
 
 
 class DeleteRowDimensionMutation(BaseMutation):
-    """Мутация для удаления строки."""
+    """Удаление строки."""
 
     class Input:
-        row_id = graphene.Int(required=True, description='Идентификатор строки')
+        row_dimension_id = graphene.ID(required=True, description='Идентификатор строки')
 
-    row_id = graphene.Int(required=True, description='Идентификатор удаленной строки')
-    index = graphene.Int(required=True, description='Измененные строки')
-    merged_cells = graphene.List(MergedCellType, required=True, description='Объединенные ячейки')
+    row_dimension_id = graphene.ID(required=True, description='Идентификатор удаленной строки')
 
     @staticmethod
     @permission_classes((IsAuthenticated,))
-    def mutate_and_get_payload(root: Any, info: ResolveInfo, row_id: int):
-        row: RowDimension = get_object_or_404(RowDimension, pk=row_id)
-        sheet: Sheet = Sheet.objects.get(pk=row.sheet_id)
-        row.delete()
-        sheet.rowdimension_set.filter(index__gt=row.index).update(index=F('index') - 1)
-        sheet.move_merged_cells(row.index, -1, True)
-        return DeleteRowDimensionMutation(row_id=row_id, index=row.index, merged_cells=sheet.mergedcell_set.all())
-
-
-class ChangeColumnDimensionPayload(DjangoCudBaseMutation, DjangoUpdateMutation):
-    """Мутация для изменения стилей колонки таблицы."""
-
-    class Meta:
-        model = ColumnDimension
-        only_fields = ('width', 'fixed', 'hidden', 'kind',)
-        required_fields = ('fixed', 'hidden', 'kind',)
-        field_types = {
-            'kind': graphene.String(description='Тип поля')
-        }
-        login_required = True
-        permissions = ('dcis.change_columndimension',)
-
-    column_dimension = graphene.Field(ColumnDimensionType, description='Измененные стили колонки таблицы')
+    def mutate_and_get_payload(root: Any, info: ResolveInfo, row_dimension_id: str):
+        return DeleteRowDimensionMutation(
+            row_dimension_id=delete_row_dimension(
+                row_dimension=get_object_or_404(RowDimension, pk=row_dimension_id)
+            )
+        )
 
 
 class ChangeCellsOptionMutation(BaseMutation):
-    """Мутация для изменения свойств ячеек:
+    """Изменение свойств ячеек:
 
-        - horizontal_align - ['left', 'center', 'right']
-        - vertical_align - ['top', 'middle', 'bottom']
-        - size - цифра от 10 до 24
         - strong - true, false
         - italic - true, false
+        - strike - true, false
         - underline - [None, 'single', 'double', 'single_accounting', 'double_accounting']
+        - horizontal_align - ['left', 'center', 'right']
+        - vertical_align - ['top', 'middle', 'bottom']
+        - size - число от 6 до 24
         - kind - [
             'n', 's', 'f', 'b', 'inlineStr', 'e', 'str', 'd', 'text', 'money',
             'bigMoney', 'fl', 'user', 'department', 'organization', 'classification'
@@ -128,50 +222,31 @@ class ChangeCellsOptionMutation(BaseMutation):
     """
 
     class Input:
-        cells_id = graphene.List(graphene.NonNull(graphene.Int), required=True, description='Идентификатор ячейки')
+        cell_ids = graphene.List(graphene.NonNull(graphene.ID), required=True, description='Идентификаторы ячеек')
         field = graphene.String(required=True, description='Идентификатор поля')
         value = graphene.String(description='Значение поля')
 
-    cells = graphene.List(CellType, description='Измененные ячейки')
-    values = graphene.List(ValueType, description='Измененные значения')
+    changed_options = graphene.List(
+        graphene.NonNull(ChangedCellOption),
+        description='Измененные свойства ячеек'
+    )
 
     @staticmethod
     @permission_classes((IsAuthenticated,))
     def mutate_and_get_payload(
         root: Any,
         info: ResolveInfo,
-        cells_id: list[int],
+        cell_ids: list[int],
         field: str,
-        value: Optional[str] = None
+        value: str | None = None
     ):
-        success, value, errors = check_cell_options(field, value)
-        if not success:
-            return ChangeCellsOptionMutation(success=success, errors=errors)
-        cells = Cell.objects.filter(pk__in=cells_id).all()
-        values: list[Value] = []
-        with transaction.atomic():
-            for cell in cells:
-                if field == 'kind':
-                    cell.save(update_fields=change_cell_kind(cell, value))
-                    values.extend(updates_values_by_cell_kind_change(cell))
-                else:
-                    setattr(cell, field, value)
-                    cell.save(update_fields=(field,))
-        return ChangeCellsOptionMutation(cells=cells, values=values)
-
-
-class UnloadFileValueArchiveMutation(BaseMutation):
-    """Выгрузка архива значения ячейки типа `Файл`."""
-
-    class Input:
-        value_id = graphene.ID(required=True, description='Идентификатор значения ячейки')
-
-    src = graphene.String(description='Ссылка на сгенерированный архив')
-
-    @staticmethod
-    @permission_classes((IsAuthenticated,))
-    def mutate_and_get_payload(root: None, info: ResolveInfo, value_id: str):
-        return UnloadFileValueArchiveMutation(src=create_file_value_archive(get_object_or_404(Value, pk=value_id)))
+        field = snakecase(field)
+        match CheckCellOptions(field, value):
+            case CheckCellOptions.Error(field, error):
+                return ChangeCellsOptionMutation(success=False, errors=[ErrorFieldType(field, [error])])
+            case CheckCellOptions.Success(value):
+                cells = Cell.objects.filter(pk__in=cell_ids).all()
+                return ChangeCellsOptionMutation(changed_options=change_cells_option(cells, field, value))
 
 
 class ChangeValueMutation(BaseMutation):
@@ -179,34 +254,33 @@ class ChangeValueMutation(BaseMutation):
 
     class Input:
         document_id = graphene.ID(required=True, description='Идентификатор документа')
-        sheet_id = graphene.Int(required=True, description='Идентификатор листа')
-        column_id = graphene.Int(required=True, description='Идентификатор колонки')
-        row_id = graphene.Int(required=True, description='Идентификатор строки')
+        sheet_id = graphene.ID(required=True, description='Идентификатор листа')
+        column_id = graphene.ID(required=True, description='Идентификатор колонки')
+        row_id = graphene.ID(required=True, description='Идентификатор строки')
         value = graphene.String(required=True, description='Значение')
 
-    value = graphene.Field(ValueType, description='Измененное значение')
+    value = graphene.String(required=True, description='Измененное значение')
+    updated_at = graphene.DateTime(required=True, description='Дата изменения')
 
     @staticmethod
     @permission_classes((IsAuthenticated,))
     def mutate_and_get_payload(
-        root: None,
+        root: Any,
         info: ResolveInfo,
         document_id: str,
-        sheet_id: int,
-        column_id: int,
-        row_id: int,
+        sheet_id: str,
+        column_id: str,
+        row_id: str,
         value: str
     ):
-        document: Document = get_object_or_404(Document, pk=from_global_id(document_id)[1])
-        sheet: Sheet = get_object_or_404(Sheet, pk=sheet_id)
-        val, _ = update_or_create_value(
-            document=document,
-            sheet=sheet,
+        result = update_or_create_value(
+            document_id=from_global_id(document_id)[1],
+            sheet_id=sheet_id,
             column_id=column_id,
             row_id=row_id,
             value=value
         )
-        return ChangeValueMutation(value=val)
+        return ChangeValueMutation(value=result.value.value, updated_at=result.updated_at)
 
 
 class ChangeFileValueMutation(BaseMutation):
@@ -214,63 +288,103 @@ class ChangeFileValueMutation(BaseMutation):
 
     class Input:
         document_id = graphene.ID(required=True, description='Идентификатор документа')
-        sheet_id = graphene.Int(required=True, description='Идентификатор листа')
-        column_id = graphene.Int(required=True, description='Идентификатор колонки')
-        row_id = graphene.Int(required=True, description='Идентификатор строки')
+        sheet_id = graphene.ID(required=True, description='Идентификатор листа')
+        column_id = graphene.ID(required=True, description='Идентификатор колонки')
+        row_id = graphene.ID(required=True, description='Идентификатор строки')
         value = graphene.String(required=True, description='Значение')
         remaining_files = graphene.List(graphene.NonNull(graphene.ID), required=True, description='Оставшиеся файлы')
         new_files = graphene.List(graphene.NonNull(Upload), required=True, description='Новые файлы')
 
-    value = graphene.Field(ValueType, description='Измененное значение')
+    value = graphene.String(required=True, description='Измененное значение')
+    updated_at = graphene.DateTime(required=True, description='Дата изменения')
     value_files = graphene.List(FileType, description='Измененные файлы')
 
     @staticmethod
     @permission_classes((IsAuthenticated,))
     def mutate_and_get_payload(
-        root: None,
+        root: Any,
         info: ResolveInfo,
         document_id: str,
-        sheet_id: int,
-        column_id: int,
-        row_id: int,
+        sheet_id: str,
+        column_id: str,
+        row_id: str,
         value: str,
-        remaining_files: list[int],
+        remaining_files: list[str],
         new_files: list[InMemoryUploadedFile]
     ):
-        document: Document = get_object_or_404(Document, pk=from_global_id(document_id)[1])
-        sheet: Sheet = get_object_or_404(Sheet, pk=sheet_id)
-        val, _ = update_or_create_file_value(
+        result = update_or_create_file_value(
             user=info.context.user,
-            document=document,
-            sheet=sheet,
+            document_id=from_global_id(document_id)[1],
+            sheet_id=sheet_id,
             column_id=column_id,
             row_id=row_id,
             value=value,
             remaining_files=[int(from_global_id(global_id)[1]) for global_id in remaining_files],
             new_files=new_files
         )
-        return ChangeFileValueMutation(value=val, value_files=get_file_value_files(val))
+        return ChangeFileValueMutation(
+            value=result.value.value,
+            updated_at=result.updated_at,
+            value_files=get_file_value_files(result.value)
+        )
+
+
+class UnloadFileValueArchiveMutation(BaseMutation):
+    """Выгрузка архива значения ячейки типа `Файл`."""
+
+    class Input:
+        document_id = graphene.ID(required=True, description='Идентификатор документа')
+        sheet_id = graphene.ID(required=True, description='Идентификатор листа')
+        column_id = graphene.ID(required=True, description='Идентификатор колонки')
+        row_id = graphene.ID(required=True, description='Идентификатор строки')
+        name = graphene.String(required=True, description='Название архива')
+
+    src = graphene.String(description='Ссылка на сгенерированный архив')
+
+    @staticmethod
+    @permission_classes((IsAuthenticated,))
+    def mutate_and_get_payload(
+        root: Any,
+        info: ResolveInfo,
+        document_id: str,
+        sheet_id: str,
+        column_id: str,
+        row_id: str,
+        name: str
+    ):
+        return UnloadFileValueArchiveMutation(
+            src=create_file_value_archive(
+                value=get_object_or_404(
+                    Value,
+                    document_id=from_global_id(document_id)[1],
+                    sheet_id=sheet_id,
+                    column_id=column_id,
+                    row_id=row_id
+                ),
+                name=name
+            )
+        )
 
 
 class SheetMutations(graphene.ObjectType):
     """Список мутаций для работы с листами документа."""
 
-    add_row_dimension = AddRowDimensionMutation.Field(required=True, description='Добавление строки')
-    delete_row_dimension = DeleteRowDimensionMutation.Field(required=True, description='Удаление строки')
+    rename_sheet = RenameSheetMutation.Field(required=True, description='Изменение названия листа')
 
-    change_column_dimension = ChangeColumnDimensionPayload.Field(
-        required=True,
-        description='Изменение стилей колонки таблицы'
-    )
+    change_column_dimension = ChangeColumnDimensionMutation.Field(required=True, description='Изменение колонки')
+
+    add_row_dimension = AddRowDimensionMutation.Field(required=True, description='Добавление строки')
+    change_row_dimension = ChangeRowDimensionMutation.Field(required=True, description='Изменение строки')
+    delete_row_dimension = DeleteRowDimensionMutation.Field(required=True, description='Удаление строки')
 
     change_cells_option = ChangeCellsOptionMutation.Field(required=True, description='Изменения опций ячейки')
 
-    unload_file_value_archive = UnloadFileValueArchiveMutation.Field(
-        required=True,
-        description='Выгрузка архива значения ячейки типа `Файл`'
-    )
     change_value = ChangeValueMutation.Field(required=True, description='Изменение значения ячейки')
     change_file_value = ChangeFileValueMutation.Field(
         required=True,
         description='Изменение значения ячейки типа `Файл`'
+    )
+    unload_file_value_archive = UnloadFileValueArchiveMutation.Field(
+        required=True,
+        description='Выгрузка архива значения ячейки типа `Файл`'
     )
