@@ -1,14 +1,15 @@
 from pathlib import PosixPath
-from typing import List, Iterator, Tuple, Dict
+from typing import Iterator, Union
 
-from openpyexcel import load_workbook, Workbook
-from openpyexcel.styles.colors import COLOR_INDEX, WHITE
-from openpyexcel.utils.cell import column_index_from_string
-from openpyexcel.worksheet.dimensions import DimensionHolder
-from openpyexcel.worksheet.merge import MergeCell
+from openpyxl import Workbook, load_workbook
+from openpyxl.cell.cell import Cell as OpenpyxlCell
+from openpyxl.styles.colors import COLOR_INDEX, WHITE
+from openpyxl.utils.cell import column_index_from_string
+from openpyxl.worksheet.dimensions import DimensionHolder
+from openpyxl.worksheet.merge import MergeCell, MergedCell as OpenpyxlMergedCell
 
 from apps.dcis.helpers.theme_to_rgb import theme_and_tint_to_rgb
-from ..models import Period, Sheet, Cell, MergedCell, RowDimension, ColumnDimension
+from ..models import Cell, ColumnDimension, MergedCell, Period, RowDimension, Sheet
 
 
 class ExcelExtractor:
@@ -17,14 +18,14 @@ class ExcelExtractor:
     def __init__(self, path: PosixPath):
         """Инициализация.
 
-        :param path - путь к файлу excel.
+        :param path - путь к файлу Excel.
         """
         self.path = path
         self.work_book = load_workbook(path)
 
     def save(self, period: Period):
         """Сохранение обработанного файла в базу данных."""
-        extract_sheets: List[Dict] = self.extract()
+        extract_sheets: list[dict] = self.extract()
         for position, extract_sheet in enumerate(extract_sheets):
             sheet = Sheet.objects.create(
                 name=extract_sheet['name'],
@@ -32,10 +33,10 @@ class ExcelExtractor:
                 period=period
             )
             # Соотношение позиции и созданных идентификаторов
-            columns_mapper: Dict[int, int] = {}
-            rows_mapper: Dict[int, int] = {}
-            columns_styles: Dict[int, Dict[str, str]] = {}
-            rows_styles: Dict[int, Dict[str, str]] = {}
+            columns_mapper: dict[int, int] = {}
+            rows_mapper: dict[int, int] = {}
+            columns_styles: dict[int, dict[str, str]] = {}
+            rows_styles: dict[int, dict[str, str]] = {}
             for cell in extract_sheet['cells']:
                 column_id: int = cell['column_id']
                 row_id: int = cell['row_id']
@@ -70,18 +71,18 @@ class ExcelExtractor:
             for merged_cell in extract_sheet['merged_cells']:
                 MergedCell.objects.create(sheet=sheet, **merged_cell)
 
-    def extract(self) -> List[Dict]:
-        """Парсинг файла эксель.
+    def extract(self) -> list[dict]:
+        """Парсинг файла Excel.
 
         Функция создает структуру данных, которая является первоначальной обработкой.
         После выделения необходимых данных можно осуществлять транзакционную запись в базу данных.
         Структура данных может использоваться для предварительной демонстрации планируемого отчета.
         """
-        sheets: List[Dict] = []
+        sheets: list[dict] = []
         for sheet in self.work_book.worksheets:
             name: str = sheet.title
-            columns_dimension: Dict[int, object] = self._parse_columns_dimension(sheet.column_dimensions)
-            rows_dimension: Dict[int, object] = self._parse_rows_dimension(sheet.row_dimensions)
+            columns_dimension: dict[int, object] = self._parse_columns_dimension(sheet.column_dimensions)
+            rows_dimension: dict[int, object] = self._parse_rows_dimension(sheet.row_dimensions)
             cells = self._parse_cells(self.work_book, sheet.rows)
             merged_cells = self._parse_merged_cells(sheet.merged_cells.ranges)
 
@@ -95,7 +96,7 @@ class ExcelExtractor:
         return sheets
 
     @staticmethod
-    def _parse_columns_dimension(holder: DimensionHolder) -> Dict:
+    def _parse_columns_dimension(holder: DimensionHolder) -> dict:
         """Парсинг имеющихся колонок."""
         return {
             column_index_from_string(col_letter): {
@@ -127,7 +128,7 @@ class ExcelExtractor:
         }
 
     @staticmethod
-    def _parse_rows_dimension(holder: DimensionHolder) -> Dict:
+    def _parse_rows_dimension(holder: DimensionHolder) -> dict:
         """Парсинг имеющихся строк."""
         return {
             row.index: {
@@ -157,42 +158,41 @@ class ExcelExtractor:
         }
 
     @staticmethod
-    def _parse_cells(wb: Workbook, rows: Iterator[Tuple[Cell]]) -> List[Dict]:
+    def _color_transform(wb, color):
+        if color and color.type == 'indexed':
+            if color.index == 64 or color.index == 65:
+                color = None
+            else:
+                color.type = 'rgb'
+                color.value = COLOR_INDEX[color.index]
+        if color and color.type == 'theme':
+            color.type = 'rgb'
+            color.value = theme_and_tint_to_rgb(wb, color.theme, color.tint)
+        return color
+
+    def _parse_cells(self, wb: Workbook, rows: Iterator[tuple[Union[OpenpyxlCell, OpenpyxlMergedCell]]]) -> list[dict]:
         """Парсинг ячеек.
 
         Переданный параметр rows представляет собой матрицу.
         Каждая строка включает в себя массив ячеек, который соотноситься с колонками.
         """
-        rows_result: List[Dict] = []
+        rows_result: list[dict] = []
         for row in rows:
             for cell in row:
-                top_color = cell.border.top.color
-                bottom_color = cell.border.bottom.color
-                left_color = cell.border.left.color
-                right_color = cell.border.right.color
-                diagonal_color = cell.border.diagonal.color
-                fill_color = cell.fill.fgColor
-                font_color = cell.font.color
-                colors = [top_color, bottom_color, left_color, right_color, diagonal_color, fill_color, font_color]
+                border_color: dict[str, Union[int, str]] = {
+                    positional: self._color_transform(wb, getattr(cell.border, positional).color)
+                    for positional in ('top', 'bottom', 'left', 'right', 'diagonal')
+                }
+                fill_color = self._color_transform(wb, cell.fill.fgColor)
+                font_color = self._color_transform(wb, cell.font.color)
 
-                for color in colors:
-                    if color and color.type == 'indexed':
-                        if color.index == 64 or color.index == 65:
-                            color = None
-                        else:
-                            color.type = 'rgb'
-                            color.value = COLOR_INDEX[color.index]
-                    if color and color.type == 'theme':
-                        color.type = 'rgb'
-                        color.value = theme_and_tint_to_rgb(wb, color.theme, color.tint)
                 # Временная заглушка
                 if (font_color and font_color.index == 1 and cell.fill.patternType is None) or \
                         (font_color and font_color.index == 1 and fill_color.value == WHITE):
                     font_color.type = 'rgb'
                     font_color.value = '00000000'
-
                 rows_result.append({
-                    'column_id': cell.col_idx,
+                    'column_id': cell.column,
                     'row_id': cell.row,
                     'kind': cell.data_type,
                     'formula': cell.value
@@ -205,7 +205,7 @@ class ExcelExtractor:
                     'size': cell.font.sz,
                     'strong': cell.font.b,
                     'italic': cell.font.i,
-                    'strike': cell.font.strike,
+                    'strike': cell.font.strike or False,
                     'underline': cell.font.u,
                     'color': f'#{font_color.value[2:]}'
                     if font_color and font_color.type == 'rgb' else '#000000',
@@ -213,31 +213,19 @@ class ExcelExtractor:
                     if cell.fill.patternType is None
                     else f'#{cell.fill.fgColor.value[2:]}',
                     'border_style': {
-                        'top': cell.border.top.style,
-                        'bottom': cell.border.bottom.style,
-                        'left': cell.border.left.style,
-                        'right': cell.border.right.style,
-                        'diagonal': cell.border.diagonal.style,
                         'diagonalDown': cell.border.diagonalDown,
-                        'diagonalUp': cell.border.diagonalUp
+                        'diagonalUp': cell.border.diagonalUp,
+                        **{p: getattr(cell.border, p).style for p in ('top', 'bottom', 'left', 'right', 'diagonal',)}
                     },
                     'border_color': {
-                        'top': f'#{top_color.value[2:]}'
-                        if top_color and top_color.type == 'rgb' else None,
-                        'bottom': f'#{bottom_color.value[2:]}'
-                        if bottom_color and bottom_color.type == 'rgb' else None,
-                        'left': f'#{left_color.value[2:]}'
-                        if left_color and left_color.type == 'rgb' else None,
-                        'right': f'#{right_color.value[2:]}'
-                        if right_color and right_color.type == 'rgb' else None,
-                        'diagonal': f'#{diagonal_color.value[2:]}'
-                        if diagonal_color and diagonal_color.type == 'rgb' else None
+                        p: f'#{c.value[2:]}' if c and c.type == 'rgb' else None
+                        for p, c in border_color.items()
                     }
                 })
         return rows_result
 
     @staticmethod
-    def _parse_merged_cells(ranges: List[MergeCell]) -> List[Dict]:
+    def _parse_merged_cells(ranges: list[MergeCell]) -> list[dict]:
         """Парсинг объединенных ячеек."""
         return [{
             'min_col': rng.min_col,
